+++ kernel/arch/generic_x86/textmode_video.c
@@ -79,9 +79,9 @@ static void textmode_scroll_one_line_up(void)
  * There is a lot of precious information about how to work with the cursor.
  */
 
-static void textmode_move_cursor(int row, int col, int color /* ignored */)
+static void textmode_move_cursor(u16 row, u16 col, int color /* ignored */)
 {
-   u16 position = (u16)((row * VIDEO_COLS) + col);
+   u16 position = (row * VIDEO_COLS) + col;
 
    // cursor LOW port to vga INDEX register
    outb(0x3D4, 0x0F);
